Build Ingredient Exposures Tables

This function builds the Ingredient Exposures Table for the algorithm to derive the antineoplastic combinations based on the RxNorm Ingredient of the drug in question.

buildIngredientExposuresTable(
  conn,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  drugExposureIngredientTable
)

Arguments

conn
Connection object to the database engine

cdmDatabaseSchema
Source CDM Drug Exposure Table schema

writeDatabaseSchema
Schema where the OncoRegimenFinder tables will be written to.

drugExposureIngredientTable
Name of the new Ingredient Exposure Table

Value

Ingredient Exposure Table in writeDatabaseSchema that contains all the original fields from the Drug Exposure Table with the addition of an Ingedient Conceot Id column that maps the drug to its RxNorm Ingredient.

Details

If the Cohort, Regimen, and Regimen Staging Tables already exist, these tables are renamed with the system date appended in the format "_YYYY_MM_DD". The renaming is skipped and the existing Cohort, Regimen and Regimen Staging Tables are dropped if the dated table already exists, indicating that the OncoRegimenFinder was already run once for that system date.

See also