Write the Vocabulary Table

The Vocabulary Table takes all the HemOnc Regimens in the OMOP Concept Table, maps the RxNorm Ingredients that comprise each Regimen, and calculates the number of unique ingredients per Regimen in an ingredient_count field and a string aggregate ingredient_combination. The ingredient_combination field combines the names of all the ingredients in a comma-separated string when the ingredient_count is greater than 2, an "and" concatenation when the ingredient_count is 2, or simply the ingredient name when ingredient_count is 1. Since a set of RxNorm Ingredients can map to more than 1 HemOnc Regimen, the table also is filtered for the shortest ingredient_combination string by character count by the concept id of the regimen, excluding abbreviations assumed to be of character counts of 5 or less.

createVocabTable(
  conn,
  writeDatabaseSchema,
  cdmDatabaseSchema,
  vocabularyTable,
  renameTable
)

Arguments

conn
PARAM_DESCRIPTION

writeDatabaseSchema
PARAM_DESCRIPTION

cdmDatabaseSchema
PARAM_DESCRIPTION

vocabularyTable
PARAM_DESCRIPTION

Value

A Vocabulary Table in the designated schema.

See also